requirements.txt
numpy==2.1.3
pandas==2.2.3
pymongo==4.10.1
python-dotenv==1.0.1
